+++ apps/ui-dev/component/status/animation/MainView.jsx
@@ -45,7 +45,7 @@ export function MainView() {
             const A = Animation[it];
             return (
               <div>
-                <A duration="2s">
+                <A duration="2s" play>
                   <ProgressCircle size="large" variant="primary" value={30} />
                 </A>
                 <div>{it}</div>