+++ apps/ui-dev/pages/main/data.jsx
@@ -1,0 +1,92 @@
+export const data = [
+  {
+    title: "@elf-framework/icon",
+    npm: "@elf-framework/icon",
+    description: "Sapa based icon component library that use material icons",
+    docLink: "https://github.com/elf-framework/icon",
+    codeLinkText: "Github",
+    codeLink: "https://github.com/elf-framework/icon",
+    type: "beta",
+  },
+  {
+    title: "@elf-framework/base-editor",
+    npm: "@elf-framework/base-editor",
+    description: "sapa based editor framework",
+    docLink: "../pages/editor/getting-started/index.html",
+    codeLinkText: "Github",
+    codeLink:
+      "https://github.com/elf-framework/editor/tree/develop/packages/base-editor",
+    type: "beta",
+  },
+  {
+    title: "@elf-framework/color",
+    npm: "@elf-framework/color",
+    description: "Elf color library",
+    docLink: "https://www.elf-framework.com",
+    codeLinkText: "Github",
+    codeLink:
+      "https://github.com/elf-framework/editor/tree/develop/packages/color",
+  },
+
+  {
+    title: "@elf-framework/ui",
+    npm: "@elf-framework/ui",
+    description: "Sapa UI component library",
+    docLink: "../pages/design-system/getting-started/index.html",
+    codeLinkText: "Github",
+    codeLink:
+      "https://github.com/elf-framework/editor/tree/develop/packages/ui",
+  },
+
+  {
+    title: "@elf-framework/design-tokens",
+    npm: "@elf-framework/design-tokens",
+    description: "Elf Design Tokens",
+    docLink: "../pages/design-system/getting-started/index.html",
+    codeLinkText: "Github",
+    codeLink:
+      "https://github.com/elf-framework/editor/tree/develop/packages/design-tokens",
+  },
+
+  {
+    title: "@elf-framework/design-system",
+    npm: "@elf-framework/design-system",
+    description: "Elf Design System",
+    docLink: "../pages/design-system/getting-started/index.html",
+    codeLinkText: "Github",
+    codeLink:
+      "https://github.com/elf-framework/editor/tree/develop/packages/design-system",
+  },
+
+  {
+    selected: true,
+    title: "@elf-framework/sapa",
+    npm: "@elf-framework/sapa",
+    description: "UI framework",
+    docLink: "../pages/sapa/getting-started/index.html",
+    codeLinkText: "Github",
+    codeLink:
+      "https://github.com/elf-framework/editor/tree/develop/packages/sapa",
+  },
+
+  {
+    title: "vite-plugin-sapa",
+    npm: "vite-plugin-sapa",
+    description: "Vite plugin to support HMR for sapa",
+    docLink: "https://www.elf-framework.com",
+    codeLinkText: "Github",
+    codeLink:
+      "https://github.com/elf-framework/editor/tree/develop/packages/vite-plugin-sapa",
+  },
+
+  {
+    title: "@elf-framework/sapa-router",
+    npm: "@elf-framework/sapa-router",
+    description: "Client router for sapa",
+    docLink: "https://www.elf-framework.com",
+    codeLinkText: "Github",
+    codeLink:
+      "https://github.com/elf-framework/editor/tree/develop/packages/sapa-router",
+    type: "beta",
+  },
+];